Create TCP-IP.C
This is not a fully-fledged TCP/IP stack. I have created a minimal TCP/IP stack to understand the functionality of networking layers. It utilizes raw sockets, bypassing the embedded TCP/IP stack of the operating system.

The code builds custom layers such as the transport layer and network layer while bypassing the OS's networking stack. The custom-built layers then hand over the data to the link (or data link) layer of the embedded stack. This layer transmits the data to the network layer, which sends the data to the network as raw packets.

This project helps in understanding the inner workings of networking layers by constructing and manipulating packets directly.